drivers: misc: aw8691: Adapt for m1882
This change contains a minimal set of changes that allows
you to use the awinic driver as before on custom ROMs.
Most likely the driver needs a little rechecking,
which will be done later.

Some things from stock were not copied. For example,
there is only single Real-Time Playback option, since
RTP firmware from Flyme is not suitable for custom ROMs.
The same situation happened with cs_press and mback support.
In fact, we only need to play effects and vibrate at
a given time, which is what works.